requirements.txt
flask
pillow
joblib